code refactor to process the large galaxies with tidal streams

+++ py/legacyhalos/coadds.py
@@ -306,6 +306,7 @@ def _custom_sky(skyargs):
     from legacypipe.reference import get_reference_sources
     from legacypipe.oneblob import get_inblob_map
 
+    log = None
     survey, brickname, brickwcs, onegal, radius_mask_arcsec, apodize, sky_annulus, ccd = skyargs
 
     im = survey.get_image_object(ccd)
@@ -315,7 +316,7 @@ def _custom_sky(skyargs):
 
     print(im, im.band, 'exptime', im.exptime, 'propid', ccd.propid,
           'seeing {:.2f}'.format(ccd.fwhm * im.pixscale), 
-          'object', getattr(ccd, 'object', None))
+          'object', getattr(ccd, 'object', None), file=log)
 
     radius_mask = np.round(radius_mask_arcsec / im.pixscale).astype('int') # [pixels]
     
@@ -339,7 +340,7 @@ def _custom_sky(skyargs):
     T = fits_table(im.merged_splineskyfn)
     I, = np.nonzero((T.expnum == im.expnum) * np.array([c.strip() == im.ccdname for c in T.ccdname]))
     if len(I) != 1:
-        print('Multiple splinesky models!')
+        print('Multiple splinesky models!', file=log)
         return 0
     splineskytable = T[I]
 
@@ -364,6 +365,9 @@ def _custom_sky(skyargs):
         inmask = (xmask**2 + ymask**2) <= 2*radius_mask**2
         outmask = (xmask**2 + ymask**2) <= 5*radius_mask**2
         skymask = (outmask*1 - inmask*1 - galmask*1) == 1
+        # There can be no sky pixels if the CCD is on the periphery.
+        if np.sum(skymask) == 0:
+            skymask = np.ones_like(img).astype(bool)
     else:
         skymask = np.ones_like(img).astype(bool)
 
@@ -392,19 +396,33 @@ def _custom_sky(skyargs):
     objmask = binary_dilation(objmask, iterations=3)
 
     skypix = ( (ivarmask*1 + refmask*1 + galmask*1 + objmask*1) == 0 ) * skymask
+    # If there are no sky pixels then use the statistics from the pipeline sky
+    # map.  For example, the algorithm here can fail in and around bright stars.
+    if np.sum(skypix) == 0:
+        print('No viable sky pixels; using pipeline sky statistics!', file=log)
+        pipesky = np.zeros_like(img)
+        tim.sky.addTo(pipesky)
+        skymean, skymed, skysig = sigma_clipped_stats(pipesky, mask=~skymask, sigma=3.0)
+        try:
+            skymode = estimate_mode(pipesky[skymask], raiseOnWarn=True).astype('f4')
+        except:
+            print('Warning: sky mode estimation failed!', file=log)
+            skymode = np.array(0.0).astype('f4')
+    else:
+        skypix = ( (ivarmask*1 + galmask*1 + objmask*1) == 0 ) * skymask
 
-    #print('ahack!!', im.expnum, im.ccdname)
-    #if im.expnum == 625736 and im.ccdname == 'S22':
-    try:
-        skymean, skymed, skysig = sigma_clipped_stats(img, mask=~skypix, sigma=3.0)
-    except:
-        skymean, skymed, skysig = 0.0, 0.0, 0.0
-    #skysig = 1.0 / np.sqrt(np.median(ivar[skypix]))
-    #skymed = np.median(img[skypix])
-    try:
-        skymode = estimate_mode(img[skypix], raiseOnWarn=True)
-    except:
-        skymode = 0.0
+        try:
+            skymean, skymed, skysig = sigma_clipped_stats(img, mask=~skypix, sigma=3.0)
+        except:
+            print('Warning: sky statistic estimates failed!', file=log)
+            skymean, skymed, skysig = 0.0, 0.0, 0.0
+        #skysig = 1.0 / np.sqrt(np.median(ivar[skypix]))
+        #skymed = np.median(img[skypix])
+        try:
+            skymode = estimate_mode(img[skypix], raiseOnWarn=True).astype('f4')
+        except:
+            print('Warning: sky mode estimation failed!', file=log)
+            skymode = np.array(0.0).astype('f4')
 
     # Build the final bit-mask image.
     #   0    = 
@@ -434,6 +452,18 @@ def _custom_sky(skyargs):
     hdr.add_record(dict(name='XCEN', value=x0-1)) # 0-indexed
     hdr.add_record(dict(name='YCEN', value=y0-1))
 
+    customsky = fits_table()
+    customsky.camera = [im.camera]
+    customsky.expnum = [im.expnum]
+    customsky.ccdname = [im.ccdname]
+    customsky.skymode = [skymode]
+    customsky.skymed = [skymed]
+    customsky.skymean = [skymean]
+    customsky.skysig = [skysig]
+    customsky.xcen = [x0 - 1] # 0-indexed
+    customsky.ycen = [y0 - 1]
+    customsky.to_np_arrays()
+
     out = dict()
     ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
     #ext = '{}-{:02d}-{}'.format(im.name, im.hdu, im.band)
@@ -441,6 +471,7 @@ def _custom_sky(skyargs):
     out['{}-image'.format(ext)] = img
     out['{}-splinesky'.format(ext)] = splineskytable
     out['{}-header'.format(ext)] = hdr
+    out['{}-customsky'.format(ext)] = customsky
     out['{}-comask'.format(ext)] = comask
     
     return out
@@ -448,7 +479,8 @@ def _custom_sky(skyargs):
 def custom_coadds(onegal, galaxy=None, survey=None, radius_mosaic=None,
                   radius_mask=None, nproc=1, pixscale=0.262, log=None,
                   apodize=False, plots=False, verbose=False, cleanup=True,
-                  write_ccddata=True, sky_annulus=True, centrals=True):
+                  write_ccddata=False, sky_annulus=True, centrals=True,
+                  doforced_phot=True):
     """Build a custom set of coadds for a single galaxy, with a custom mask and sky
     model.
 
@@ -551,47 +583,54 @@ def custom_coadds(onegal, galaxy=None, survey=None, radius_mosaic=None,
     [sky.update(res) for res in result]
     del result
 
-    if write_ccddata:
-        # Write out the "coadd" mask.
-        cokeys = [key for key in sky.keys() if 'comask' in key]
-        _comask = np.array([sky[cokey] for cokey in cokeys])
+    # Write out the "coadd" mask.
+    cokeys = [key for key in sky.keys() if 'comask' in key]
+    _comask = np.array([sky[cokey] for cokey in cokeys])
 
-        comask = np.bitwise_or.reduce(_comask, axis=0)
-        hdr = fitsio.FITSHDR()
-        brickwcs.add_to_header(hdr)
-        hdr.delete('IMAGEW')
-        hdr.delete('IMAGEH')
+    comask = np.bitwise_or.reduce(_comask, axis=0)
+    hdr = fitsio.FITSHDR()
+    brickwcs.add_to_header(hdr)
+    hdr.delete('IMAGEW')
+    hdr.delete('IMAGEH')
 
-        maskfile = os.path.join(survey.output_dir, '{}-custom-mask-grz.fits.gz'.format(galaxy))
-        fitsio.write(maskfile, comask, header=hdr, clobber=True)
-        print('Writing {}'.format(maskfile), flush=True, file=log)
-        del comask
+    maskfile = os.path.join(survey.output_dir, '{}-custom-mask-grz.fits.gz'.format(galaxy))
+    fitsio.write(maskfile, comask, header=hdr, clobber=True)
+    print('Writing {}'.format(maskfile), flush=True, file=log)
+    del comask
 
-        skyfile = os.path.join(survey.output_dir, '{}-pipeline-sky.fits'.format(galaxy))
-        print('Writing {}'.format(skyfile), flush=True, file=log)
-        if os.path.isfile(skyfile):
-            os.remove(skyfile)
-        for ii, ccd in enumerate(survey.ccds):
-            im = survey.get_image_object(ccd)
-            ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
-            sky['{}-splinesky'.format(ext)].write_to(skyfile, append=ii>0, extname=ext)
+    skyfile = os.path.join(survey.output_dir, '{}-pipeline-sky.fits'.format(galaxy))
+    print('Writing {}'.format(skyfile), flush=True, file=log)
+    if os.path.isfile(skyfile):
+        os.remove(skyfile)
+    for ii, ccd in enumerate(survey.ccds):
+        im = survey.get_image_object(ccd)
+        ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
+        sky['{}-splinesky'.format(ext)].write_to(skyfile, append=ii>0, extname=ext)
+
+    skyfile = os.path.join(survey.output_dir, '{}-custom-sky.fits'.format(galaxy))
+    print('Writing {}'.format(skyfile), flush=True, file=log)
+    if os.path.isfile(skyfile):
+        os.remove(skyfile)
+    for ii, ccd in enumerate(survey.ccds):
+        im = survey.get_image_object(ccd)
+        ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
+        sky['{}-customsky'.format(ext)].write_to(skyfile, append=ii>0, extname=ext)
         
-        # Write out separate CCD-level files with the images/data, individual masks
-        # (converted to unsigned integer), and the pipeline/splinesky binary FITS
-        # tables.
-        print('NEED to write out the custom sky values as a FITS table!')
-        pdb.set_trace()
-        if False:
-            ccdfile = os.path.join(survey.output_dir, '{}-custom-ccdmask-grz.fits.gz'.format(galaxy))
-            print('Writing {}'.format(ccdfile), flush=True, file=log)
-            if os.path.isfile(ccdfile):
-                os.remove(ccdfile)
-            with fitsio.FITS(ccdfile, 'rw') as ff:
-                for ii, ccd in enumerate(survey.ccds):
-                    im = survey.get_image_object(ccd)
-                    ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
-                    hdr = sky['{}-header'.format(ext)]
-                    ff.write(sky['{}-mask'.format(ext)], extname=ext, header=hdr)
+    # Optionally write out separate CCD-level files with the images/data and
+    # individual masks (converted to unsigned integer).  These are still pretty
+    # big and I'm not sure we will ever need them.  Keep the code here for
+    # legacy value but don't write out.
+    if write_ccddata:
+        ccdfile = os.path.join(survey.output_dir, '{}-custom-ccdmask-grz.fits.gz'.format(galaxy))
+        print('Writing {}'.format(ccdfile), flush=True, file=log)
+        if os.path.isfile(ccdfile):
+            os.remove(ccdfile)
+        with fitsio.FITS(ccdfile, 'rw') as ff:
+            for ii, ccd in enumerate(survey.ccds):
+                im = survey.get_image_object(ccd)
+                ext = '{}-{}-{}'.format(im.camera, im.expnum, im.ccdname.lower().strip())
+                hdr = sky['{}-header'.format(ext)]
+                ff.write(sky['{}-mask'.format(ext)], extname=ext, header=hdr)
 
         # These are the actual images, which results in a giant file.  Keeping
         # the code here for legacy purposes but I'm not sure we should ever
@@ -644,26 +683,29 @@ def custom_coadds(onegal, galaxy=None, survey=None, radius_mosaic=None,
     pipeline_srcs = read_fits_catalog(pipeline_cat, fluxPrefix='')
     custom_srcs = [src.copy() for src in pipeline_srcs]
 
-    print('Performing forced photometry on the custom sky-subtracted CCDs.', flush=True, file=log)
-    forcedflx = mp.map(_forced_phot, [(custom_tims, custom_srcs, band) for band in bands])
-    if False:
-        forcedflx = []
-        for band in bands:
-            t0 = time.time()
-            forcedflx.append(forced_phot(custom_tims, custom_srcs, band))
-            print('  band {} took {:.3f} sec'.format(band, time.time()-t0), flush=True, file=log)
-
-    # Populate the new custom catalog and write out.
     custom_cat = pipeline_cat.copy()
-    for band, flux, ivar in forcedflx:
-        custom_cat.set('flux_{}'.format(band), flux.astype('f4'))
-        custom_cat.set('flux_ivar_{}'.format(band), ivar.astype('f4'))
+    if doforced_phot:
+        t0 = time.time()
+        print('Performing forced photometry on the custom sky-subtracted CCDs.', flush=True, file=log)
+        forcedflx = mp.map(_forced_phot, [(custom_tims, custom_srcs, band) for band in bands])
+        #forcedflx = []
+        #for band in bands:
+        #    forcedflx.append(forced_phot(custom_tims, custom_srcs, band))
+        print('  Total time for forced photometry = {:.3f} min'.format(time.time()-t0), flush=True, file=log)
+
+        # Populate the new custom catalog and write out.
+        for band, flux, ivar in forcedflx:
+            custom_cat.set('flux_{}'.format(band), flux.astype('f4'))
+            custom_cat.set('flux_ivar_{}'.format(band), ivar.astype('f4'))
+
+        tractorfile = os.path.join(survey.output_dir, '{}-custom-tractor.fits'.format(galaxy))
+        if os.path.isfile(tractorfile):
+            os.remove(tractorfile)
+        custom_cat.writeto(tractorfile)
+        print('Wrote {} sources to {}'.format(len(custom_cat), tractorfile), flush=True, file=log)
         
-    tractorfile = os.path.join(survey.output_dir, '{}-custom-tractor.fits'.format(galaxy))
-    if os.path.isfile(tractorfile):
-        os.remove(tractorfile)
-    custom_cat.writeto(tractorfile)
-    print('Wrote {} sources to {}'.format(len(custom_cat), tractorfile), flush=True, file=log)
+    else:
+        print('Skipping forced photometry on the custom sky-subtracted CCDs.', flush=True, file=log)
         
     custom_srcs = read_fits_catalog(custom_cat, fluxPrefix='')
